move components to component folder
